move AddEvents to screens
